zGALIv1 spend fix v2:
As we have already zGALI from v1 activation block with payload of v2 in the
chain, it is easier to align v1 with v2 activation. It is possible to move
it to the past because signatures and serials are included. In future the
code can be simplified due to this. Added new chain parameter
'nBlockFirstGoodChecksum' which indicates first valid Zerocoin v2 checksums
in the chain.