Dynamic zPoS v2:
In Galilel v3.1.0 it was a technology preview feature because we still
use zGALI denominator 1, so it means in Phase 9 we need to generate
6 * 1 zGALI which is very CPU intensive and probability to generate an
orphan block is higher since a public coin staker can solve the block
later but distribute it to the chain earlier.

In Galilel v3.1.1 we auto resolve best denomination structure. In our
case it means the following:

- v3.1.0: 6 * 1 zGALI (6 mints, each requires CPU time)
- v3.1.1: 1 * 1 zGALI + 1 * 5 zGALI (2 mints, lower probability to
generate an orphan block)

The biggest benefit for blockchain developers is that it makes coin
forks with zPoS much easier as auto resolve of best denomination
structure will even work for coins which generate very high amounts
per block.